review: use dedicated files for renewal
- rename tests to follow Go conventions
- use certcrypto.ParsePEMCertificate
